Move pedigree functions to a separate file
